+++ scripts/run_notebook_pack.sh
@@ -13,6 +13,11 @@ NOTEBOOKS=(
   notebooks/04_governed_text_to_sql_agent.ipynb
   notebooks/05_agent_red_team_evaluation_harness.ipynb
   notebooks/06_ci_gate_for_data_ai_changes.ipynb
+  notebooks/07_governed_rag_embedding_ingestion_gate.ipynb
+  notebooks/08_snowflake_cortex_agent_tool_preflight.ipynb
+  notebooks/09_openai_agents_tool_guard_pattern.ipynb
+  notebooks/10_human_approval_packet_for_conditional_export.ipynb
+  notebooks/11_llamaindex_governed_retrieval_pattern.ipynb
 )
 
 if [[ "${METATATE_EXAMPLES_MODE}" == "live" ]]; then